services/backend/requirements_integrated.txt
# Integrated Chonost System Requirements v2.1.0
# ระบบที่ผนวกรวมทุกฟีเจอร์เข้าด้วยกัน

# Core Flask
Flask==2.3.3
Flask-CORS>=4.0.1
Flask-SQLAlchemy==3.0.5

# AI & ML Libraries
openai==1.3.0
sentence-transformers==2.2.2
transformers>=4.40.0
torch>=2.7.1-rc1
numpy==1.24.3
pandas==2.0.3
scikit-learn>=1.4.0

# Vector Database
faiss-cpu==1.7.4

# Async Support
aiohttp>=3.9.0
aiofiles==23.2.1
asyncio-mqtt==0.13.0

# Database
sqlite3

# Utilities
python-dotenv==1.0.0
requests>=2.32.0
Pillow>=10.2.0
matplotlib==3.7.2
seaborn==0.12.2

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0

# Optional: GPU Support (uncomment if needed)
# faiss-gpu==1.7.4
# torch>=2.7.1-rc1+cu118 --index-url https://download.pytorch.org/whl/cu118
